telegram-bot-vercel-deployment/vercel-deployment/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-telegram-bot==20.7
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
cryptography==41.0.7
beautifulsoup4==4.12.2

